test a function that returns a string
A test written for such a function should not be too specific so
it does not break easily.if the returned string was changed from
'welcome dela' to 'hi dela', for example, the test should not break.
Testing for the returned string to contain a particular string or
is a good way of ensuring that the test does not break esily.